% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information_criteria.R
\name{AICc}
\alias{AICc}
\title{Calculate AICc}
\usage{
AICc(log_likelihood, n, k)
}
\arguments{
\item{log_likelihood}{The log-likelihood of the model}

\item{n}{The number of observations}

\item{k}{The number of estimated parameters}
}
\description{
Calculates the Akaike Information Criterion corrected for small sample sizes (AICc).
}
