% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information_criteria.R
\name{BIC2}
\alias{BIC2}
\title{Bayesian Information Criterion (BIC)}
\usage{
BIC2(log_likelihood, n, k)
}
\arguments{
\item{log_likelihood}{The log-likelihood of the model}

\item{n}{The number of observations}

\item{k}{The number of estimated parameters}
}
\description{
Calculates the Bayesian Information Criterion (BIC).
}
