% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{bray1}
\alias{bray1}
\title{Bray-Curtis dissimilarity components}
\usage{
bray1(x, y)
}
\arguments{
\item{x}{numeric vector of the first sample to be compared}

\item{y}{numeric vector of the second sample to be compared}
}
\value{
A named numeric vector with two elements:
\itemize{
\item num - numerator: sum of absolute differences between x and y
\item den - denominator: sum of x and y values
}
}
\description{
Calculate the numerator and denominator components of the Bray-Curtis dissimilarity index.
The Bray-Curtis dissimilarity is calculated as \eqn{\frac{\sum_{i} |x_i - y_i|}{\sum_{i} (x_i + y_i)}}
}
