% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gdmm.R
\name{coef_to_pred}
\alias{coef_to_pred}
\title{Convert model coefficients to predictions}
\usage{
coef_to_pred(
  object,
  intercept,
  beta,
  beta_p,
  lambda,
  u,
  new_W,
  new_X,
  new_X_pair,
  new_re,
  D,
  n,
  component,
  scale_uniq,
  type,
  sigma
)
}
\arguments{
\item{object}{Model object containing fitted model components}

\item{intercept}{Numeric, model intercept}

\item{beta}{Numeric vector, coefficients for dissimilarity component}

\item{beta_p}{Numeric vector, coefficients for dissimilarity component (pairwise)}

\item{lambda}{Numeric vector, coefficients for uniqueness component}

\item{u}{Numeric vector, random effect values}

\item{new_W}{Matrix/data.frame, new data for uniqueness component}

\item{new_X}{Matrix/data.frame, new data for dissimilarity component}

\item{new_X_pair}{Matrix/data.frame, new data for dissimilarity component (pairwise)}

\item{new_re}{Matrix/data.frame, new random effect data}

\item{D}{Matrix, pairwise combination indices}

\item{n}{Integer, number of observations}

\item{component}{Character, "dissimilarity" or "uniqueness"}

\item{scale_uniq}{Logical, whether to scale uniqueness values}

\item{type}{Character, "response" or "link" scale}

\item{sigma}{Numeric, random effect standard deviations}
}
\value{
Numeric vector of predictions (dissimilarities or uniqueness values)
}
\description{
Internal function that transforms model parameters and new data into
dissimilarity or uniqueness predictions.
}
\keyword{internal}
