% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdmmTMB-package.R
\docType{package}
\name{gdmmTMB-package}
\alias{gdmmTMB}
\alias{gdmmTMB-package}
\title{gdmmTMB: Generalised Dissimilarity Mixed Models in R Using TMB}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
