% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse logit}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{A numeric value or vector}
}
\value{
A (transformed) numeric value or vector
}
\description{
Performs inverse of logit: exp(x) / (1 + exp(x))
}
