% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_y_df.R
\name{make_y_df}
\alias{make_y_df}
\title{Format community data}
\usage{
make_y_df(
  com = NULL,
  D = NULL,
  id = NULL,
  method = "bray",
  num_den = FALSE,
  trans = NULL,
  drop_empty_rows = FALSE,
  drop_empty_cols = TRUE,
  binary = FALSE,
  check_trans = TRUE,
  na.rm = FALSE,
  na.replace = NULL
)
}
\arguments{
\item{com}{community matrix (sites x species)}

\item{D}{Optional. Numeric matrix with two columns}

\item{id}{Optional. Unique id for each sample}

\item{method}{Method used to calculate community dissimilarity between pairs of sites. Either a distance metric supported by vegan, 'abcd' to obtain each component of the binary contingecy matrix, 'decomp1' or 'decomp2' for Pierre Legendre's or Andres Baselga's turnover and nestedness decomposition}

\item{num_den}{For some dissimilarity indices, whether function should return the numerator and denominator in two different columns}

\item{trans}{Transformation to be applied to raw data before calculating distance matrix. Either a function or 'binary' to transform to presence-absence)}

\item{drop_empty_rows}{Indicates whether empty rows (i.e., sites containing 0s only) should be removed}

\item{drop_empty_cols}{Indicates whether empty columns (i.e., species without observations) should be removed}

\item{binary}{Indicates whether abundance data should be transform into presence absence}

\item{check_trans}{Check for non-finite values in transformation}

\item{na.rm}{Remove rows with NA?}

\item{na.replace}{Replace NAs with specific value}
}
\value{
A data frame with the following columns
\itemize{
\item s1, s2  -  A combination of sites (rows).
\item Additional columns indicating the dissimilarity between sites.
}
}
\description{
Formats species x site table into a long data frame of community dissimilarities
}
\keyword{internal}
