% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gdmm.R
\name{print.bbgdmm}
\alias{print.bbgdmm}
\title{Print method for bbgdmm objects}
\usage{
\method{print}{bbgdmm}(x, ...)
}
\arguments{
\item{x}{A fitted model object of class \code{bbgdmm}}

\item{...}{Additional parameters (not used)}
}
\description{
Prints basic information from a fitted bbgdmm object
}
\seealso{
\code{\link{print.gdmm}} \code{\link{gdmm}}
}
