% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gdmm.R
\name{print.gdmm}
\alias{print.gdmm}
\title{Print method for gdmm objects}
\usage{
\method{print}{gdmm}(x, ...)
}
\arguments{
\item{x}{A fitted model object of class \code{gdmm}}

\item{...}{Additional parameters (not used)}
}
\description{
Prints basic information from a fitted gdmm object
}
\seealso{
\code{\link{print.bbgdmm}} \code{\link{gdmm}}
}
