% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{print_title}
\alias{print_title}
\title{Nice titles for summary tables}
\usage{
print_title(text, symb = "-", width = getOption("width"))
}
\arguments{
\item{text}{Title to display}

\item{symb}{Symbol used above and below the title}

\item{width}{Size of the console. Defaults to current width.}
}
\value{
A \code{character}
}
\description{
Nice titles for summary tables
}
\examples{
print_title('hello world', '=')
}
