% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{print_title2}
\alias{print_title2}
\title{Centered title with decorated side}
\usage{
print_title2(text, symb = "-", width = getOption("width"))
}
\arguments{
\item{text}{Title to display}

\item{symb}{Symbol used left and right of the title}

\item{width}{Size of the console. Defaults to current width.}
}
\value{
A \code{character} string (invisibly returned)
}
\description{
Centered title with decorated side
}
\examples{
print_title2(' hello world ', ':')
}
