% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{scale01}
\alias{scale01}
\title{Scale a numeric vector 0,1}
\usage{
scale01(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of the same length as \code{x}, with values in the interval [0, 1].
}
\description{
Takes a numeric vector and linearly rescales it so that its minimum becomes 0 and its maximum becomes 1.
}
