% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{scale_dist}
\alias{scale_dist}
\title{Scale a vector into an specified}
\usage{
scale_dist(x, new = c(0.01, 0.99))
}
\arguments{
\item{x}{A numeric vector}

\item{new}{Numeric scalar: the upper and lower bound of the target interval. Default is [\code{0.01}, \code{0.99}].}
}
\value{
A numeric vector of the same length as \code{x}, with values in the interval [\code{min(new)}, \code{max(new)}].
}
\description{
Given a numeric vector \code{x}, this function re-scales \code{x} into the interval [\code{min}, \code{max}].
}
