% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_summary.R
\name{summary.bbgdmm}
\alias{summary.bbgdmm}
\title{Summary method for bbgdmm objects}
\usage{
\method{summary}{bbgdmm}(obj, quantiles = c(0.025, 0.5, 0.975), null_value = 0)
}
\arguments{
\item{obj}{An object of class \code{bbgdmm}.}

\item{quantiles}{A numeric vector of quantiles to display.}

\item{null_value}{Value used for pseudo p-value tests.}
}
\value{
An object of class \code{summary.bbgdmm}.
}
\description{
S3 method for summarizing bbgdmm model fits.
}
