% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.gdmm}
\alias{summary.gdmm}
\title{Summary method for gdmm objects}
\usage{
\method{summary}{gdmm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{gdmm}.}

\item{...}{additional arguments (not used)}
}
\value{
An object of class \code{summary.gdmm}.
}
\description{
S3 method for summarizing gdmm model fits.
}
