% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diss_gradient.R
\name{diss_gradient}
\alias{diss_gradient}
\title{Predicted Effects of Environmental Gradients on Dissimilarity}
\usage{
diss_gradient(
  m,
  var = "all",
  n = 100,
  CI = TRUE,
  n_sim = NULL,
  CI_quant = c(0.95)
)
}
\arguments{
\item{m}{A fitted model object of class \code{gdmm} or \code{bbgdmm}.}

\item{var}{Character vector of variable names to evaluate. Use \code{"all"} (default) to use all predictors in the model.}

\item{n}{Integer. Number of points to evaluate along each gradient. Defaults to 100.}

\item{CI}{Logical. Whether to compute confidence intervals. Default is \code{TRUE}.}

\item{n_sim}{Number of draws used to obtain CI. If \code{NULL}, uses \code{n_boot} for \code{bbgdmm} objects or 1000 for \code{gdmm}.}

\item{CI_quant}{Numeric vector specifying the quantile coverage (e.g., \code{0.95} for 95\% CI). Default is \code{0.95}.}
}
\value{
A named list of data frames, one for each variable in \code{var}. Each data frame contains:
\itemize{
\item \code{var}: Variable names.
\item \code{f_x}: Predicted value after fitted non-linear transformation (e.g., monotonic I-splines).
\item \code{x}: The values of the focal gradient.
\item \code{CI lower / CI upper}: (Optional) Lower and upper bounds of the confidence interval.
}
}
\description{
Computes the predicted effect of one or more covariates on compositional dissimilarity
from a fitted \code{gdmm} or \code{bbgdmm} model. Optionally, returns uncertainty bands (credible or confidence intervals)
using either bootstrapped samples or parametric bootstrapping.
}
\details{
The function generates predicted dissimilarities along a gradient of each specified predictor,
while holding all other predictors constant. If \code{CI = TRUE},
it uses bootstrap samples (for \code{bbgdmm}) or simulates from the joint covariance matrix (for \code{gdmm})
to construct confidence or credible intervals.

This approach is useful for visualizing directional compositional changes along individual
environmental gradients.
}
