% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_def.R
\name{isp}
\alias{isp}
\title{Ispline Basis with Degree 2 by Default}
\usage{
isp(x, degree = 2, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{degree}{Spline degree; defaults to 2 here (instead of 3).}

\item{...}{additional parameters passed to \code{splines2::isp}}
}
\description{
A thin wrapper around \code{splines2::isp()}. Uses \code{degree = 2}
unless otherwise specified. All other arguments are passed through.
}
