% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gdmm.R
\name{print.bbgdmm}
\alias{print.bbgdmm}
\title{Print method for bbgdmm objects}
\usage{
\method{print}{bbgdmm}(m, ...)
}
\arguments{
\item{m}{A fitted model object of class \code{bbgdmm}}
}
\description{
Prints basic information from a fitted bbgdmm object
}
\seealso{
\code{\link{print.gdmm}} \code{\link{gdmm}}
}
