% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_summary.R
\name{summary.gdmm}
\alias{summary.gdmm}
\title{Summary method for gdmm objects}
\usage{
\method{summary}{gdmm}(obj)
}
\arguments{
\item{obj}{An object of class \code{gdmm}.}
}
\value{
An object of class \code{summary.gdmm}.
}
\description{
S3 method for summarizing gdmm model fits.
}
